requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
aio-pika==9.3.1
pydantic==2.5.0
alembic==1.13.0
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
